requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
boto3==1.40.6
botocore==1.40.6
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
cryptography==46.0.3
Django==5.2.1
django-cleanup==9.0.0
django-environ==0.12.0
django-jazzmin==3.0.1
django-simple-history==3.10.1
django-storages==1.14.6
django-user-sessions==2.0.0
djangorestframework==3.16.1
et_xmlfile==2.0.0
html5lib==1.1
idna==3.11
jmespath==1.0.1
lxml==6.0.2
openpyxl==3.1.5
oscrypto==1.3.0
packaging==25.0
pillow==11.2.1
pipdeptree==2.30.0
psycopg==3.2.12
psycopg-binary==3.2.12
pycparser==2.23
pyHanko==0.31.0
pyhanko-certvalidator==0.29.0
pypdf==6.2.0
pypdfium2==5.1.0
python-bidi==0.6.7
python-dateutil==2.9.0.post0
PyYAML==6.0.3
qrcode==8.2
reportlab==4.4.4
requests==2.32.5
s3transfer==0.13.1
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
webencodings==0.5.1
whitenoise==6.9.0
xhtml2pdf==0.2.17
